requirements_copy.txt
accelerate==0.31.0
aioboto3==13.1.0
aiobotocore==2.13.1
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.7.4
charset-normalizer
click==8.1.7
cohere==5.5.8
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
deeplake==3.9.11
deepspeed==0.14.4
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
eval_type_backport==0.2.0
exceptiongroup==1.2.1
fastavro==1.9.5
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
hjson==3.1.0
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humbug==0.3.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.12
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langsmith==0.1.84
libdeeplake==0.0.133
llama-cloud==0.0.6
llama-index==0.10.50
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.50
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.29
llama-index-readers-github==0.1.9
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-deeplake==0.1.4
llama-parse==0.4.6
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
neural_compressor==2.6
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
nvidia-ml-py==12.555.43
onnx==1.16.1
openai==1.35.7
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pathos==0.3.2
pdf2image==1.17.0
peft==0.11.1
pillow==10.2.0
platformdirs==4.2.2
pox==0.3.4
ppft==1.7.6.8
prettytable==3.10.0
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycocotools==2.0.8
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ragas==0.1.9
regex==2024.5.15
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
safetensors==0.4.3
schema==0.7.7
scikit-learn==1.5.1
scipy==1.13.1
sentry-sdk==2.8.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
striprtf==0.0.26
sympy==1.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.42.3
trl==0.9.4
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
urllib3==1.26.19
wandb==0.17.3
wcwidth==0.2.13
wikipedia==1.4.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
